requirements.txt
bottle==0.12.13
certifi==2017.11.5
chardet==3.0.4
Django==1.8.6
django-braces==1.8.1
django-embed-video==1.0.0
django-memcache-status==1.1
djangorestframework==3.2.3
first==2.0.1
idna==2.6
pew==0.1.26
Pillow==2.9.0
pip-tools==1.9.0
pipenv==7.3.1
psycopg2==2.7.3.2
python3-memcached==1.51
pythonz-bd==1.11.4
pytz==2017.3
requests==2.18.4
resumable-urlretrieve==0.1.5
semver==2.7.8
six==1.11.0
urllib3==1.22
virtualenv-clone==0.2.6
web-pdb==1.3.2
